refactor: extract MCP-config builders, dispatch on Auth
